+++ src/parser/bmp/messages/mod.rs
@@ -2,16 +2,14 @@
 //!
 //! <https://datatracker.ietf.org/doc/html/rfc7854>
 
-pub use headers::{
-    parse_bmp_common_header, parse_per_peer_header, BmpCommonHeader, BmpMsgType, BmpPerPeerHeader,
-};
-pub use initiation_message::{parse_initiation_message, InitiationMessage};
-pub use peer_down_notification::{parse_peer_down_notification, PeerDownNotification};
-pub use peer_up_notification::{parse_peer_up_notification, PeerUpNotification};
-pub use route_mirroring::{parse_route_mirroring, RouteMirroring};
-pub use route_monitoring::{parse_route_monitoring, RouteMonitoring};
-pub use stats_report::{parse_stats_report, StatsReport};
-pub use termination_message::{parse_termination_message, TerminationMessage};
+pub use headers::*;
+pub use initiation_message::*;
+pub use peer_down_notification::*;
+pub use peer_up_notification::*;
+pub use route_mirroring::*;
+pub use route_monitoring::*;
+pub use stats_report::*;
+pub use termination_message::*;
 
 pub(crate) mod headers;
 pub(crate) mod initiation_message;
@@ -22,15 +20,17 @@ pub(crate) mod route_monitoring;
 pub(crate) mod stats_report;
 pub(crate) mod termination_message;
 
-#[derive(Debug)]
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct BmpMessage {
     pub common_header: BmpCommonHeader,
     pub per_peer_header: Option<BmpPerPeerHeader>,
-    pub message_body: MessageBody,
+    pub message_body: BmpMessageBody,
 }
 
-#[derive(Debug)]
-pub enum MessageBody {
+#[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub enum BmpMessageBody {
     PeerUpNotification(PeerUpNotification),
     PeerDownNotification(PeerDownNotification),
     InitiationMessage(InitiationMessage),